ros/requirements.txt
pyyaml
numpy
opencv-python
opencv-contrib-python
